rewrote the mock server to start in the background
