+++ test/mock/src/start.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+#
+# This file is part of Machine Economy Zilliqa Dapp Project.
+#
+# Copyright (c) 2019 Rustiq Technology Ltd & Well Bred Software Ltd
+# MIT License
+
+node src/start.js &
